feat: change constructor of user, implement Rider class, implement Driver class
Updated the constructor for userclass. Removed contact information from the constructor as it makes more sense to add the contact information after an instance of the user has been created. This can be done via a mandatory view that appears after the user creates their account. In this view, they add their contact information.

Added code to the shell of Rider & Driver class