fix: fixed layout of DriverSignupActivity
Fixed the layout of the driver sign up activity. It was completely broken before.

Update ids of each of the text views and removed the text from the "text" field as it is only suppoed to be present in the hint field.

The ids will be used to connect this to datainput later so this was important.