feat: added comments and required methods for Driver & ContactInformation class
Added comments Driver & Contact Information class
Added methods to Driver class to add ratings. Previous setter method only allowed us to add an arraylist of ratings. Now, we can add an individual rating by calling the addRating(Rating rating) method.

Updated the constructor of the class so an empty ArrayList of ratings is initialized. This will avoid any runtime errors that we may encounter if we try to use the addRating method.