feat: adds method to mydatabase class, updates DriverHomeActivity and RiderModeActivity
Added a method to retrieve a drive request by its unique ID in the myDatabase class
Added setter method for driverId to the Driverequest entity class
Updated riderModeActivity and set the reset password button to invisible as it currently does not work. Requires logic.
Added logic to FragmentDisplayDriveRequestInfo so if a driver clicks accept, it updates the status of the DriveRequest in the databse. It also sets the driverID field of the request.